requirements.txt
streamlit==1.33.0
langchain==0.1.17
langchain-community==0.0.38
langchain-core==0.1.53
langchain-groq==0.0.1

sentence-transformers

faiss-cpu

pypdf
pdfplumber
python-dotenv
